Works correctly now, split up the loading to seperate Header
