+++ 0_Nested_Loop/NestedLoopUtils.hpp
@@ -24,7 +24,6 @@
 #include <span>
 #include <functional>
 #include <string>
-#include <iomanip>
 
 
 //==--------------------------------------------------------------------==//
@@ -286,143 +285,6 @@ inline ResultRelation createResultTuple(const CastRelation& cast, const TitleRel
     result.roleId = cast.roleId;
 
     return result;
-}
-
-std::streampos getFileSize(const std::string& filepath) {
-    std::ifstream file(filepath, std::ifstream::ate | std::ifstream::binary);
-    return file.tellg();
-}
-
-void findChunkBoundaries(const std::string& filepath, std::streampos& start, std::streampos& end) {
-    std::cout << std::this_thread::get_id() <<"::findChunkBoundaries: " << filepath << "\n";
-    std::ifstream file(filepath);
-    file.seekg(start);
-    std::string dummy;
-    if(start == 0) {
-        std::getline(file, dummy);  // When we start we need to skip the first csv line. Afters "start" is always the previous "end" so aligned to newline in file
-    }
-    std::cout << std::this_thread::get_id() << "::findChunkBoundaries: line starts with \'" << static_cast<char>(file.peek()) << "\'\n";
-    start = file.tellg();
-    file.seekg(end);
-    std::getline(file, dummy);
-    end = file.tellg();
-    if(end == start) {
-        // File is probably very small, so we initial start and stop where in the same line
-        std::getline(file, dummy);
-        end = file.tellg();
-    }
-    std::cout << std::this_thread::get_id() << "::findChunkBoundaries: line ends (after) \'" << static_cast<char>(file.peek())<< "\'" << std::endl;
-    file.close();
-}
-
-template <typename Relation>
-void processChunk(const std::string& chunk, std::istringstream& raw, std::vector<Relation>& data, std::mutex& m_data) {
-    std::string line;
-    Relation record;
-
-}
-
-
-template<typename Relation>
-std::vector<Relation> threadedLoad(const std::string& filepath, const size_t& bufferSize, const size_t numberOfTuples = SIZE_MAX) {
-    std::vector<Relation> data;
-    std::queue<std::string> chunks;
-    std::mutex m_data;
-    std::mutex m_queue;
-    std::mutex m_cout;
-    std::condition_variable cv_queue;
-    std::atomic<bool> stop(false);
-
-    auto const workerThread = [&] {
-        {
-            std::lock_guard<std::mutex> lock(m_cout);
-            std::cout << std::this_thread::get_id() << ": Started\n";
-        }
-        std::string chunk;
-        std::istringstream raw;
-        std::string line;
-        Relation record;
-        while(!stop || !chunks.empty()) {
-            std::unique_lock<std::mutex> lock(m_queue);
-            cv_queue.wait(lock, [&] {return stop || chunks.empty(); });
-            if(!chunks.empty()) { // As lock is created when stop is set, chunks might be empty when thread awakes
-                chunk = chunks.front();
-                chunks.pop();
-                lock.unlock();
-                assert(!chunk.empty());
-                raw.str(std::move(chunk));
-                data.reserve( // Maybe this helps avoid unnecessary allocation, needs testing to see if benefical
-                    std::ranges::count_if(chunk.begin(), chunk.end(), [&](const char c) {return c == '\n';}) + data.size()
-                );
-                while(!std::getline(raw, line).eof()) {
-                    if(parseLine(line, record)) {
-                        std::lock_guard<std::mutex> lock(m_data);
-                        data.emplace_back(std::move(record));
-                    } else {
-                        std::lock_guard<std::mutex> lock(m_cout);
-                        std::cerr << std::this_thread::get_id() << ": Error parsing line " << line << '\n';
-                    }
-                }
-            }
-        }
-    };
-    std::vector<std::thread> threads(std::thread::hardware_concurrency());
-    for (size_t i = 0; i < threads.size(); ++i) {
-        threads[i] = std::thread(workerThread);
-    }
-
-
-    std::ifstream file(filepath, std::ios::in);
-    if(!file) {
-        m_cout.lock();
-        std::cerr << std::this_thread::get_id() << ": Could not open " << filepath << '\n';
-        m_cout.unlock();
-        stop = true;
-        cv_queue.notify_all();
-        for(auto& t : threads) {
-            t.join();
-        }
-        return data;
-    }
-    file.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
-    std::string readBuffer = "";
-    readBuffer.resize(BLOCK_SIZE);
-    std::string leftovers = "";
-    while(file) {
-        file.read(&readBuffer[0], readBuffer.size());
-        if(!leftovers.empty()) {
-            readBuffer.insert(
-                readBuffer.begin(), leftovers.begin(), leftovers.end()
-            );
-            leftovers.clear();
-        }
-        const auto newline = readBuffer.rfind('\n');
-        if(newline != std::string::npos) {
-            leftovers.assign(readBuffer.begin() + newline + 1, readBuffer.end());
-            readBuffer.resize(newline + 1);
-        }
-        {
-            std::lock_guard<std::mutex> lock(m_queue);
-            chunks.emplace(readBuffer);
-            cv_queue.notify_one();
-        }
-
-    }
-    stop.store(true);
-    cv_queue.notify_all();
-    for(auto& t: threads) {
-        t.join();
-    }
-    return data;
-}
-
-
-inline std::vector<TitleRelation> threadedLoadTitleRelation(const std::string& filepath, const size_t& bufferSize, const size_t numberOfTuples = SIZE_MAX) {
-    return threadedLoad<TitleRelation>(filepath, bufferSize, numberOfTuples);
-}
-
-inline std::vector<CastRelation> threadedLoadCastRelation(const std::string& filepath, const size_t& bufferSize, const size_t numberOfFlags = SIZE_MAX) {
-    return threadedLoad<CastRelation>(filepath, bufferSize, numberOfFlags);
 }
